refactor: move the getProtocol and getSpecType methods from WSDL to Feature
